Refactor PromptChain and PromptRail with PromptBase
